requirements.txt
langchain
openai
llama-index-core
pydantic
python-dotenv
requests
gcsfs